Support different chunk versions

+++ NbtTools/Entities/Trading/TradeComponent.cs
@@ -6,10 +6,10 @@ namespace NbtTools.Entities.Trading
     public class TradeComponent
     {
         public Item Item { get; private set; }
-        public byte Quantity { get; private set; }
+        public int Quantity { get; private set; }
         public ICollection<Enchantment> Enchantments { get; private set; }
 
-        public TradeComponent(Item item, byte quantity, ICollection<Enchantment> enchantments)
+        public TradeComponent(Item item, int quantity, ICollection<Enchantment> enchantments)
         {
             Item = item;
             Quantity = quantity;
